refactor: refactor directory structure of manage
