+++ src/pages/manage/Navigator.tsx
