--- src/pages/manage/index.css
@@ -1,3 +1,0 @@
-body {
-    font-family: 'Pretendard';
-}
